refactor(server): extract the chat backend to server/chat.py (ADR 0023 phase 2)
Third backend extraction. Moves the LangGraph turn loop out of the composition
root into server/chat.py:
- chat (non-streaming, Gradio + OpenAI-compat) + _chat_langgraph
- _chat_langgraph_stream (streaming, the A2A handler's driver)
- _run_turn_stream (the shared astream_events loop)
- tool-preview / interrupt shaping (_coerce_tool_*, _interrupt_payload)
- slash-command parsing + execution (_parse_*/_run_parsed_* for workflows + subagents)
- _setup_required_message

chat.py imports only neutral modules (runtime.state, graph.output_format) plus
function-local imports — nothing from server/__init__ — so unlike a2a.py there's
no `from server import` and zero cycle risk. server/__init__.py re-exports every
name and drops ~645 lines; _main's OpenAI-compat + A2A wiring calls them as
before.

Note: the submodule server.chat and the re-exported function `chat` share a
name, so the `server.chat` *attribute* resolves to the function. Every real
access pattern still works — `from server import chat` (the function, used by 11
test files) and `from server.chat import X` / importlib (the module, via the
import system). Only `import server.chat` attribute-chaining would pick up the
function, and nothing does that.

Verified: 1000 tests green; live smoke via `python -m server` — a non-streaming
OpenAI-compat turn (chat -> _chat_langgraph) and a streaming A2A
SendStreamingMessage turn (_chat_langgraph_stream) both round-tripped
(TASK_STATE_COMPLETED + cost-v1).
